+++ lerobot/common/robot_devices/control_utils.py
@@ -36,79 +36,12 @@ from lerobot.common.policies.pretrained import PreTrainedPolicy
 from lerobot.common.robot_devices.robots.utils import Robot
 from lerobot.common.robot_devices.utils import busy_wait
 from lerobot.common.utils.utils import get_safe_torch_device, has_method
-
-import mink
-from robot_descriptions.loaders.mujoco import load_robot_description
-import numpy as np
-import pytorch3d.transforms as transforms
-from tqdm import tqdm
-from torch import pi
-
-ALOHA_MODEL = load_robot_description("aloha_mj_description")
-ALOHA_CONFIGURATION = mink.Configuration(ALOHA_MODEL)
-
-
-def map_real2sim(Q):
-    """
-    The real robot joints and the sim robot don't map exactly to each other.
-    Some joints are offset, some joints rotate the opposite direction.
-    This mapping converts real robot joint angles (in radians) to the sim version.
-
-    sim = real*sign + offset
-    """
-    # Set gripper fingers to 0, we don't care about them for IK
-    sign = torch.tensor([-1, -1, 1, 1, 1, 1, 0, 0,    
-                      -1, -1, 1, 1, 1, 1, 0, 0])
-    offset = torch.tensor([pi/2, 0, -pi/2, 0, 0, 0, 0, 0,
-                       pi/2, 0, -pi/2, 0, 0, 0, 0, 0])
-    Q = sign*Q + offset
-
-    # We handle the shoulder joint separately, x*-1 + np.pi/2 brings it close but just outside joint limits for some reason....
-    # Remap this joint range using real observed min/max and sim min/max
-    real_min, real_max = -3.59, -0.23
-    sim_min, sim_max = -1.85, 1.26 
-    Q[1] = (Q[1] - real_min)*((sim_max-sim_min)/(real_max-real_min)) + sim_min
-    Q[9] = (Q[9] - real_min)*((sim_max-sim_min)/(real_max-real_min)) + sim_min
-
-    return Q
-
-def forward_kinematics(ALOHA_CONFIGURATION, real_joints):
-    # Mapping from LeRobot (real) to robot_descriptions (sim)
-    # Check LeRobot joint names in lerobot/common/robot_devices/robots/configs.py
-    # Check robot_descriptions joint names with `print([model.joint(i).name for i in range(model.njnt)])`
-    Q = torch.deg2rad(torch.tensor(
-        [
-            real_joints[0],
-            real_joints[1],
-            real_joints[3],
-            real_joints[5],
-            real_joints[6],
-            real_joints[7],
-            0,
-            0,
-
-            real_joints[9],
-            real_joints[10],
-            real_joints[12],
-            real_joints[14],
-            real_joints[15],
-            real_joints[16],
-            0,
-            0,
-        ]
-    ))
-    Q = map_real2sim(Q)
-    ALOHA_CONFIGURATION.update(Q)
-    eef_pose_se3 = ALOHA_CONFIGURATION.get_transform_frame_to_world("right/gripper", "site")
-    rot_6d, trans = transforms.matrix_to_rotation_6d(torch.from_numpy(eef_pose_se3.as_matrix()[None, :3, :3])).squeeze(), torch.from_numpy(eef_pose_se3.as_matrix()[:3,3])
-    eef_pose = torch.cat([rot_6d, trans], axis=0)
-    return eef_pose, eef_pose_se3
+from lerobot.common.utils.aloha_utils import ALOHA_CONFIGURATION, forward_kinematics
 
 def add_eef_pose(real_joints):
     eef_pose, eef_pose_se3 = forward_kinematics(ALOHA_CONFIGURATION, real_joints)
     eef_pose = torch.cat([eef_pose, real_joints[-1][None]], axis=0).float()
     return eef_pose
-
 
 def log_control_info(robot: Robot, dt_s, episode_index=None, frame_index=None, fps=None):
     log_items = []
